unify MAD e_misc.c
